@@ -1,13 +1,35 @@
 package com.example.manishvishwakarma.searchablespinnermine;
 
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
+import android.widget.Spinner;
+
+import androidx.appcompat.app.AppCompatActivity;
 
 public class MainActivity extends AppCompatActivity {
 
+    Spinner PartSpinner;
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
+    /*    PartSpinner=(Spinner)findViewById(R.id.spinnerPart);
+
+        ArrayAdapter<String> adapter = new ArrayAdapter<>(this, android.R.layout.simple_spinner_dropdown_item, getResources().getStringArray(R.array.colleges));
+        PartSpinner.setAdapter(adapter);
+        PartSpinner.setOnItemSelectedListener(new AdapterView.OnItemSelectedListener() {
+
+            @Override
+            public void onItemSelected(AdapterView<?> adapter, View v, int position, long id) {
+                // On selecting a spinner item
+                ((TextView) adapter.getChildAt(0)).setTextColor(Color.BLACK);
+//                Part_Id = PartId.get(position);
+
+            }
+
+            @Override
+            public void onNothingSelected(AdapterView<?> arg0) {
+                // TODO Auto-generated method stub
+            }
+        });*/
     }
 }
